Refactor: place background patterns in a container
This has been done to prevent side-scrolling caused by the overflow of
the background patterns (circles) while maintaining the ability to
scroll down if more content is added to the page.

This had been previously done by placing the main and footer elements in
a container, but fewer lines of CSS are required if the background
patterns are placed in a container instead.